debugging

+++ Agents/DPG.py
@@ -65,7 +65,7 @@ class DPGAgent(torch.nn.Module):
 
                 # Explore phase
                 if self.step < self.explore_steps and self.training:
-                    action = action.float().uniform_(-1, 1)
+                    action = action.uniform_(-1, 1)
 
             if self.discrete:
                 action = torch.argmax(action, -1)
